pyproject.toml
[tool.poetry]
name = "setitup"
version = "0.1.0"
description = "Sets up the development environment to your likings."
authors = ["Peter Yuen"]

[tool.poetry.dependencies]
python = "^3.10"
click = "^8.0.3"
shutils = "^0.1.0"
termcolor = "^1.1.0"
toml = "^0.10.2"
flatdict = "^4.0.1"
pydantic = "^1.8.2"
Pygments = "^2.10.0"
PyYAML = "^6.0"
i-lazy = "^0.1.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
autopep8 = "^1.6.0"

[tool.poetry.scripts]
setup_env = "setitup.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
